src/requirements.txt
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
astroid==2.12.13
async-timeout==4.0.2
attrs==22.1.0
bidict==0.22.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.2.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.2.1
exceptiongroup==1.0.4
fastapi==0.87.0
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
jedi==0.18.1
kombu==5.2.4
lazy-object-proxy==1.8.0
mccabe==0.7.0
motor==3.1.1
msgpack==1.0.4
multidict==6.0.4
packaging==21.3
parso==0.8.3
platformdirs==2.5.4
pluggy==1.0.0
prompt-toolkit==3.0.32
pydantic==1.10.2
PyJWT==2.6.0
pylint==2.15.6
pymongo==4.3.2
pynvim==0.4.3
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.3
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2022.6
redis==4.3.4
requests==2.30.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
spilo==0.0.19
starlette==0.21.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
uvicorn==0.19.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
